# backend/swagger.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    Full-stack Project & Task Management Platform API.
    
    ## Authentication
    Most endpoints require a JWT access token in the `Authorization: Bearer <token>` header.
    
    ## Roles
    - **ADMIN** — Full access to all resources
    - **PROJECT_MANAGER** — Manage own projects and their tasks
    - **TEAM_MEMBER** — Manage assigned tasks only
  version: 1.0.0
  contact:
    name: TaskFlow Team
servers:
  - url: http://localhost:5000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15min expiry)
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, PROJECT_MANAGER, TEAM_MEMBER]
        isActive:
          type: boolean
        isVerified:
          type: boolean
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [ADMIN, PROJECT_MANAGER, TEAM_MEMBER]
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Admin only
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                role:
                  type: string
                  enum: [ADMIN, PROJECT_MANAGER, TEAM_MEMBER]
                isActive:
                  type: boolean
      responses:
        '200':
          description: User updated
        '404':
          description: User not found
    delete:
      tags: [Users]
      summary: Delete (deactivate) user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deactivated
        '404':
          description: User not found
...
  /activity/user/{userId}:
    get:
      tags: [Activity]
      summary: Get user activity history
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User activity log
...